requirements.txt
colorcet==0.9.1
matplotlib==2.1.2
numpy==1.14.0
pycuda==2017.1.1
Pillow==5.0.0
PyQt5==5.10
